requirements.txt
asgiref==3.8.1
cachetools==5.4.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
cryptography==43.0.0
dj-rest-auth==6.0.0
Django==5.0.8
django-allauth==64.0.0
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
google-auth==2.33.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
httplib2==0.22.0
idna==3.7
oauthlib==3.2.2
packaging==24.2
pillow==11.0.0
psycopg2-binary==2.9.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.9.0
pyparsing==3.1.2
PyPDF2==3.0.1
pytesseract==0.3.13
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.2
